LangGraph-usercase/pyproject.toml
[project]
name = "langgraph-usercase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ibm>=1.0.3",
    "langchain[anthropic]>=1.2.9",
    "langgraph>=1.0.8",
    "langsmith>=0.6.9",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
]
